Rename output format handler using kebab case
