+++ src/output_format_handler.py
@@ -1,38 +1,15 @@
-from enum import Enum
-import re
-from bs4 import BeautifulSoup
-from markdownify import markdownify as md
+import importlib.util
+import pathlib
 
+spec = importlib.util.spec_from_file_location(
+    __name__,
+    pathlib.Path(__file__).with_name('output-format-handler.py')
+)
+_module = importlib.util.module_from_spec(spec)
+spec.loader.exec_module(_module)
 
-class OutputFormat(Enum):
-    """Supported output formats for scraped content."""
-
-    MARKDOWN = "markdown"
-    TEXT = "text"
-    HTML = "html"
-
-
-def to_markdown(html: str) -> str:
-    """Convert sanitized HTML to Markdown."""
-    return md(html)
-
-
-def to_text(html: str) -> str:
-    """Convert sanitized HTML to plain text."""
-    soup = BeautifulSoup(html, "html.parser")
-    text = soup.get_text(separator="\n", strip=True)
-    return re.sub(r"\n\s*\n", "\n\n", text).strip()
-
-
-def to_html(html: str) -> str:
-    """Return sanitized HTML as-is."""
-    return html
-
-
-def format_content(html: str, output_format: OutputFormat) -> str:
-    """Return content in the desired format."""
-    if output_format == OutputFormat.TEXT:
-        return to_text(html)
-    if output_format == OutputFormat.HTML:
-        return to_html(html)
-    return to_markdown(html)
+OutputFormat = _module.OutputFormat
+format_content = _module.format_content
+to_markdown = _module.to_markdown
+to_text = _module.to_text
+to_html = _module.to_html
